test: 스토리북 Global Styles 및 theme 적용
